Fix setup of default log_format in PuppetLink.configuration when it is empty.
PuppetLink#log_format checks that configuration.log_format == '', but in
fact it receives nil. So nil is not equal ''. No default configuration
is set as result and we get an error in PuppetLint#format_message about
no method % for nil.

This commit fixes it by checking that value is nil? or empty? which is
proper way to determine absent value for configuration.log_format.