Resource: fix nested ensure error.
If an ensure statement is defined within a hash inside a resource, then
ensure_param_name_token will end up being set to nil. This is because
first_param_comma_token jumps to the comma that ends the hash.

This is valid puppet syntax, so the linter shouldn't error out.  With
puppet default arguments, it is normal to pass a hash to a class, which
are applied with the * operator, and ensure could be one of the
arguments. For that reason, moving the ensure out of the inner hash will
change the semantics of the code. Instead, just raise a NoFix and the
user will see the warning.