Correctly lex non-keyword type tokens in interpolation
There's a conflict in the current lexer that where the `type` keyword
token (`:type = :TYPE, :value => 'type'`) conflicts with the known type
name tokens (e.g. `:type => :TYPE, :value => 'Integer'`).

The proper solution to this would be to use a different token type for
these two use cases, but that is a breaking change as the token types
are considered public API for semver purposes. Until the next major
release where this can be introduced, this fix will do.